requirements.txt
# RAG-QA System Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-generativeai>=0.3.0
faiss-cpu>=1.7.4
pymupdf>=1.23.0
slowapi>=0.1.9
python-dotenv>=1.0.0
aiofiles>=23.2.1
numpy>=1.24.0

# Testing
pytest>=7.4.0
httpx>=0.25.0
pytest-asyncio>=0.21.0
